# contracts/openapi/zen70-gateway-full.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZEN70 API",
    "version": "1.58.0"
  },
  "paths": {
...
    "/api/v1/settings/flags": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "List Flags",
        "operationId": "list_flags_api_v1_settings_flags_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response List Flags Api V1 Settings Flags Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/settings/flags/{key}": {
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Toggle Flag",
        "operationId": "toggle_flag_api_v1_settings_flags__key__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Toggle Flag Api V1 Settings Flags  Key  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/config": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "List Config",
        "operationId": "list_config_api_v1_settings_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response List Config Api V1 Settings Config Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/settings/config/{key}": {
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Update Config",
        "operationId": "update_config_api_v1_settings_config__key__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Config Api V1 Settings Config  Key  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/schema": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Settings Schema",
        "operationId": "get_settings_schema_api_v1_settings_schema_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsSchemaResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SettingsSchemaField": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "input": {
            "type": "string",
            "title": "Input",
            "default": "text"
          },
          "editable": {
            "type": "boolean",
            "title": "Editable",
            "default": false
          },
          "save_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Save Path"
          },
          "placeholder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placeholder"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label"
        ],
        "title": "SettingsSchemaField"
      },
      "SettingsSchemaResponse": {
        "properties": {
          "product": {
            "type": "string",
            "title": "Product"
          },
          "profile": {
            "type": "string",
            "title": "Profile"
          },
          "runtime_profile": {
            "type": "string",
            "title": "Runtime Profile"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/SettingsSchemaSection"
            },
            "type": "array",
            "title": "Sections"
          }
        },
        "type": "object",
        "required": [
          "product",
          "profile",
          "runtime_profile"
        ],
        "title": "SettingsSchemaResponse"
      },
      "SettingsSchemaSection": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/SettingsSchemaField"
            },
            "type": "array",
            "title": "Fields"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label"
        ],
        "title": "SettingsSchemaSection"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}